added moveset_alternatives_rby.txt
Contains several pokemon who appear to have multiple viable movesets. In most cases, the highest moveset is the one that I reasonably guess will have the highest overall power. If I was unable to determine a best moveset definitively, the pokemon is identified as a "suspect" at the top. Further verification will be needed for these pokemon.